[Master] feat : add uuid to boargame model

+++ boardgames/apis/boardgame.py
@@ -2,6 +2,7 @@ import uuid
 from dao.models import Boardgame
 from flask_restplus import Namespace, Resource, fields
 from flask import Response
+import uuid
 
 from dao.db import db
 
@@ -10,7 +11,7 @@ api = Namespace('boardgames', description='Boardgames related operations')
 
 boardgame = api.model(
     'Boardgame',
-    {'id': fields.String(description='The technical id'),
+    {'id': fields.String(description='The id', attribute="boardgame_id"),
      'name': fields.String(description='The name', required=True),
      'publisher': fields.String(description='The color', required=True),
      'minplayer': fields.Integer(description='The minimum required number of player'),
@@ -45,11 +46,15 @@ class BoardgamesHandler(Resource):
     @ api.expect(boardgame, validate=True)
     def post(self):
         body = api.payload
+        print(body)
         del body["id"]
+        body["boardgame_id"] = str(uuid.uuid4())
+
         db.session.add(Boardgame(**body))
         db.session.commit()
         db.session.close()
-        return Response(status=201)
+
+        return {"id": body["boardgame_id"]}, 201
 
 
 @ api.route('/<id>')
@@ -61,16 +66,27 @@ class BoardgameHandler(Resource):
     @ api.doc(responses={200: 'OK'})
     @ api.doc(responses={404: 'Not Found'})
     @ api.doc(responses={500: 'Internal Error'})
-    @ api.marshal_with(boardgame, code=200)
+    @ api.marshal_with(boardgame)
     def get(self, id):
-        return Boardgame.query.get_or_404(id)
+        if not is_uuid4(id):
+            return api.abort(404)
+
+        return Boardgame.query.filter_by(boardgame_id=id).first_or_404()
+
+        # using if using id primarykey
+        # return Boardgame.query.get_or_404(id)
 
     @ api.doc(params={'id': 'Technical boardgame id'})
     @ api.doc(responses={204: 'Deleted'})
     @ api.doc(responses={404: 'Not Found'})
     @ api.doc(responses={500: 'Internal Error'})
     def delete(self, id):
-        boardgame_to_delete = Boardgame.query.get_or_404(id)
+        if not is_uuid4(id):
+            return api.abort(404)
+
+        boardgame_to_delete = Boardgame.query.filter_by(
+            boardgame_id=id).first_or_404()
+
         db.session.delete(boardgame_to_delete)
         db.session.commit()
         db.session.close()
@@ -83,10 +99,16 @@ class BoardgameHandler(Resource):
     @ api.doc(responses={500: 'Internal Error'})
     @ api.expect(boardgame, validate=True)
     def put(self, id):
-        Boardgame.query.get_or_404(id)
-        body = api.payload
 
-        Boardgame.query.filter_by(id=id).update({**body})
+        if not is_uuid4(id):
+            return api.abort(404)
+
+        Boardgame.query.filter_by(boardgame_id=id).first_or_404()
+
+        body = api.payload
+        del body["id"]
+
+        Boardgame.query.filter_by(boardgame_id=id).update({**body})
         db.session.commit()
         db.session.close()
 
@@ -101,23 +123,23 @@ class BoardgameHandler(Resource):
     def patch(self, id):
         body = api.payload
 
-        Boardgame.query.get_or_404(id)
+        Boardgame.query.filter_by(
+            boardgame_id=id).first_or_404()
+
+        del body["id"]
         body = api.payload
 
-        Boardgame.query.filter_by(id=id).update({**body})
+        Boardgame.query.filter_by(boardgame_id=id).update({**body})
         db.session.commit()
         db.session.close()
 
         return Response(status=200)
 
 
-# def getFilteredCats(*, search_criteria):
-
-#     criteria_dict = {}
-
-#     for criteria in search_criteria:
-#         criteria_value = request.args.get(criteria)
-#         if criteria_value:
-#             criteria_dict[criteria] = criteria_value
-
-#     return Cat.objects(**criteria_dict) if search_criteria else Cat.objects()
+# custom validation to improve
+def is_uuid4(uuid_string, version=4):
+    try:
+        uid = uuid.UUID(uuid_string, version=version)
+        return uid.hex == uuid_string.replace('-', '')
+    except ValueError:
+        return False
